Separate SumTree impl from DQNSolver
