+++ snake/solver/dqn.py
@@ -10,6 +10,7 @@ import numpy as np
 import tensorflow as tf
 
 from snake.base import Direc, PointType
+from snake.util.sumtree import SumTree
 from snake.solver.base import BaseSolver
 
 
@@ -23,25 +24,11 @@ def _log(*msgs):
     print("[%s]" % datetime.now().strftime("%Y-%m-%d %H:%M:%S"), msg_str)
 
 
-class _SumTree:
+class _SumTree(SumTree):
 
     PATH_TREE = os.path.join(_DIR_LOG, "sumtree-tree-%d.npy")
     PATH_DATA = os.path.join(_DIR_LOG, "sumtree-data-%d.npy")
     PATH_VAR = os.path.join(_DIR_LOG, "sumtree-var-%d.json")
-
-    def __init__(self, capacity):
-        self.__capacity = capacity
-        self.__tree = np.zeros(2 * self.__capacity - 1)
-        self.__data = np.zeros(self.__capacity, dtype=object)
-        self.__data_idx = 0
-
-    @property
-    def tree(self):
-        return self.__tree
-
-    @property
-    def data(self):
-        return self.__data
 
     def save(self, steps):
         np.save(_SumTree.PATH_TREE % steps, self.__tree)
@@ -61,44 +48,6 @@ class _SumTree:
         self.__data_idx = var["data_idx"]
         _log("sum tree loaded | tree shape: {} | data shape: {} | capacity: {} | data_idx: {}"
              .format(self.__tree.shape, self.__data.shape, self.__capacity, self.__data_idx))
-
-    def sum(self):
-        return self.__tree[0]
-
-    def insert(self, data, priority):
-        self.__data[self.__data_idx] = data
-        tree_idx = self.__data_idx + self.__capacity - 1
-        self.update(tree_idx, priority)
-        self.__data_idx += 1
-        if self.__data_idx >= self.__capacity:
-            self.__data_idx = 0
-
-    def update(self, tree_idx, priority):
-        delta = priority - self.__tree[tree_idx]
-        self.__tree[tree_idx] = priority
-        while tree_idx != 0:
-            tree_idx = (tree_idx - 1) // 2  # Get parent
-            self.__tree[tree_idx] += delta
-
-    def retrieve(self, val):
-        parent = 0
-        while True:
-            left = 2 * parent + 1
-            right = left + 1
-            if left >= len(self.__tree):
-                tree_idx = parent
-                break
-            else:
-                if val <= self.__tree[left]:
-                    parent = left
-                else:
-                    val -= self.__tree[left]
-                    parent = right
-        data_idx = tree_idx - self.__capacity + 1
-        return tree_idx, self.__tree[tree_idx], self.__data[data_idx]
-
-    def leaves(self):
-        return self.__tree[-self.__capacity:]
 
 
 class _Memory:
