+++ snake/util/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+"""Snake utils package."""
